move Elements::Footing to Elements::RC::Footing::Base
